Remove legacy monolithic user handler
