+++ app/texts_data.json
@@ -86,7 +86,8 @@
     "calculation_error": "❌ <b>Ошибка при расчете КБЖУ</b>\n\nПопробуй начать заново через /start",
     "profile_error": "❌ <b>Ошибка отображения профиля</b>\n\nПопробуйте еще раз.",
     "temp_error": "❌ <b>Временная ошибка</b>\n\nПопробуйте еще раз через несколько секунд.",
-    "general_error": "❌ Произошла ошибка. Попробуйте еще раз."
+    "general_error": "❌ Произошла ошибка. Попробуйте еще раз.",
+    "processing": "Ошибка обработки данных, попробуйте еще раз."
   },
   "questions": {
     "gender": "👤 <b>Давай знакомиться!</b>\n\nВыбери свой пол:",
@@ -102,7 +103,13 @@
   },
   "profile": {
     "no_data": "👤 <b>Твой профиль</b>\n\n❌ У тебя еще нет рассчитанных КБЖУ\nНажми \"Рассчитать КБЖУ\" чтобы получить персональные данные!",
-    "template": "👤 <b>Твой профиль</b>\n\n📊 <b>Данные:</b>\n• Пол: {gender_icon} {gender_text}\n• Возраст: {age} лет\n• Рост: {height} см\n• Вес: {weight} кг\n• Активность: {activity_text}\n• Цель: {goal_text}\n\n🔥 <b>Твоя норма КБЖУ:</b>\n• Калории: {calories} ккал/день\n• Белки: {proteins} г\n• Жиры: {fats} г\n• Углеводы: {carbs} г\n\n📅 Рассчитано: {calc_date}"
+    "template": "👤 <b>Твой профиль</b>\n\n📊 <b>Данные:</b>\n• Пол: {gender_icon} {gender_text}\n• Возраст: {age} лет\n• Рост: {height} см\n• Вес: {weight} кг\n• Активность: {activity_text}\n• Цель: {goal_text}\n\n🔥 <b>Твоя норма КБЖУ:</b>\n• Калории: {calories} ккал/день\n• Белки: {proteins} г\n• Жиры: {fats} г\n• Углеводы: {carbs} г\n\n📅 Рассчитано: {calc_date}",
+    "not_specified": "не указано",
+    "date_format": "%d.%m.%Y",
+    "gender_male_icon": "👨",
+    "gender_female_icon": "👩",
+    "gender_male": "Мужской",
+    "gender_female": "Женский"
   },
   "kbju_start": {
     "title": "Начало расчета КБЖУ",
@@ -150,12 +157,50 @@
     "write_lead": "💬 Написать лиду",
     "nav_prev": "◀️ Предыдущий",
     "nav_next": "▶️ Следующий",
-    "back": "🔙 Назад"
+    "back": "🔙 Назад",
+    "leads_refresh": "🔄 Обновить",
+    "confirm_yes": "✅ Да",
+    "confirm_no": "✖️ Нет"
   },
   "admin": {
     "welcome": "👋 Привет! Админ-панель открыта.",
     "no_hot_leads": "📈 <b>Горячих лидов нет</b>",
     "counter": "📊 {current} из {total}",
-    "lead_card": "{priority_icon} <b>Лид #{index}</b>\n\n👤 <b>{first_name}</b>\n🆔 ID: <code>{tg_id}</code>\n💬 Username: {username_link}\n🏆 Приоритет: <b>{priority_score}</b>\n📈 Статус: <code>{funnel_status}</code>\n\n📊 <b>КБЖУ данные</b>\n• Пол: {gender_text}\n• Возраст: {age_text}\n• Рост: {height_text}\n• Вес: {weight_text}\n• Активность: {activity_text}\n• Цель: {goal_text}\n• Калории: {calories_text}\n\n🎯 <b>Направление:</b> {priority_label}\n🕓 <i>Обновлен: {updated_at}</i>"
+    "lead_card": "{priority_icon} <b>Лид #{index}</b>\n\n👤 <b>{first_name}</b>\n🆔 ID: <code>{tg_id}</code>\n💬 Username: {username_link}\n🏆 Приоритет: <b>{priority_score}</b>\n📈 Статус: <code>{funnel_status}</code>\n\n📊 <b>КБЖУ данные</b>\n• Пол: {gender_text}\n• Возраст: {age_text}\n• Рост: {height_text}\n• Вес: {weight_text}\n• Активность: {activity_text}\n• Цель: {goal_text}\n• Калории: {calories_text}\n\n🎯 <b>Направление:</b> {priority_label}\n🕓 <i>Обновлен: {updated_at}</i>",
+    "leads": {
+      "left_title": "Лид покинул бот",
+      "no_permission": "Недостаточно прав.",
+      "no_permission_alert": "Недостаточно прав",
+      "invalid_request": "Некорректный запрос",
+      "missing_message": "Нет сообщения",
+      "load_error": "Не удалось загрузить список.",
+      "load_error_alert": "Ошибка загрузки",
+      "empty_list": "Список пуст.",
+      "delete_prompt": "Вы уверены, что хотите удалить лида <code>{tg_id}</code>?",
+      "delete_failed": "⚠️ Не удалось удалить лида. Попробуйте позже.",
+      "delete_failed_alert": "Не удалось",
+      "delete_success": "✅ Лид <code>{tg_id}</code> удалён.\nЕсли карточка всё ещё отображается, нажмите «🔄 Обновить».",
+      "delete_success_alert": "Удалено",
+      "delete_cancelled_alert": "Отменено",
+      "pager_label": "Лиды: страница {page} из {total_pages} • Всего: {total_count} • Фильтр: {window_label}",
+      "pager_window_all": "все",
+      "pager_window_today": "за 24 часа",
+      "pager_window_7d": "за 7 дней"
+    }
+  },
+  "fallbacks": {
+    "welcome_plain": "Добро пожаловать!",
+    "default_first_name": "Пользователь",
+    "username_unknown": "не указан"
+  },
+  "debug": {
+    "ping_response": "pong"
+  },
+  "contact": {
+    "forward_success": "Ваше сообщение отправлено администратору.",
+    "forward_failure": "Не удалось отправить сообщение админу, попробуйте позже."
+  },
+  "defaults": {
+    "channel_username": "@fitness_channel"
   }
 }
